+++ swarm/storage/localstore/gc.go
