+++ swarm/storage/localstore/mode_get.go
