+++ swarm/storage/localstore/mode_set.go
