+++ swarm/storage/localstore/mode_set_test.go
