+++ swarm/storage/localstore/retrieval_index_test.go
